requirements.txt
Flask
wheel
Pillow
face_recognition